requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
openpyxl
httpx
python-multipart
pydantic
aiofiles
apscheduler
slowapi
